// entidad-web/entidad-web/src/app/@data/model/cabeceraMaestra.ts
export class CabeceraMaestra {
    estadoRegistro: string;
    usuarioCreacion: string;
    fechaCreacion: string;
    usuarioModificacion: string;
    fechaModificacion: string;
    maeCabeceraId: number;
    codigoCabecera: string;
    descripcion: string;
    descripcionCorta: string;
    soloServir: string;
    listMaestraDetalles: string;
    listaMaestraDetalleEntidads: string;
}